requirements.txt
pytest
requests
PyYAML
allure-pytest
python-dotenv
